Test Camera and some Structure change and rename
aa
