+++ Assets/Scripts/PlayerCameraRotateController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9b50454619a144f4ca8d3972e7b10172
+guid: f64e4011c007f914e923488547029f4f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
